+++ backend/src/main/java/com/carffeine/carffeine/controller/dto/ChargeStationSimpleResponse.java
@@ -1,4 +1,4 @@
-package com.carffeine.carffeine.dto;
+package com.carffeine.carffeine.controller.dto;
 
 import java.math.BigDecimal;
 import java.util.List;
