+++ backend/src/main/java/com/carffeine/carffeine/controller/dto/ChargeStationsSimpleResponse.java
@@ -1,4 +1,4 @@
-package com.carffeine.carffeine.dto;
+package com.carffeine.carffeine.controller.dto;
 
 import com.carffeine.carffeine.domain.ChargeStation;
