+++ backend/src/main/java/com/carffeine/carffeine/service/ChargerStationService.java
@@ -4,7 +4,7 @@ import com.carffeine.carffeine.domain.ChargeStation;
 import com.carffeine.carffeine.domain.ChargeStationRepository;
 import com.carffeine.carffeine.domain.Latitude;
 import com.carffeine.carffeine.domain.Longitude;
-import com.carffeine.carffeine.dto.CoordinateRequest;
+import com.carffeine.carffeine.service.dto.CoordinateRequest;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
