+++ backend/src/main/java/com/carffeine/carffeine/service/ScrapperService.java
@@ -3,7 +3,7 @@ package com.carffeine.carffeine.service;
 import com.carffeine.carffeine.domain.ChargeStation;
 import com.carffeine.carffeine.domain.ChargeStationRepository;
 import com.carffeine.carffeine.domain.Charger;
-import com.carffeine.carffeine.dto.ChargeStationRequest;
+import com.carffeine.carffeine.service.dto.ChargeStationRequest;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
