+++ backend/src/main/java/com/carffeine/carffeine/service/dto/ChargeStationRequest.java
@@ -1,4 +1,4 @@
-package com.carffeine.carffeine.dto;
+package com.carffeine.carffeine.service.dto;
 
 public record ChargeStationRequest(
         int numOfRows,
