+++ backend/src/main/java/com/carffeine/carffeine/service/dto/ChargeStationsInfoRequest.java
@@ -1,4 +1,4 @@
-package com.carffeine.carffeine.dto;
+package com.carffeine.carffeine.service.dto;
 
 import com.carffeine.carffeine.domain.ChargeStation;
