+++ backend/src/main/java/com/carffeine/carffeine/service/dto/CoordinateRequest.java
@@ -1,4 +1,4 @@
-package com.carffeine.carffeine.dto;
+package com.carffeine.carffeine.service.dto;
 
 import java.math.BigDecimal;
