+++ backend/src/test/java/com/carffeine/carffeine/controller/ChargerStationControllerTest.java
@@ -1,9 +1,9 @@
 package com.carffeine.carffeine.controller;
 
 import com.carffeine.carffeine.domain.ChargeStation;
-import com.carffeine.carffeine.dto.CoordinateRequest;
 import com.carffeine.carffeine.fixture.ChargeStationFixture;
 import com.carffeine.carffeine.service.ChargerStationService;
+import com.carffeine.carffeine.service.dto.CoordinateRequest;
 import com.fasterxml.jackson.databind.ObjectMapper;
 import org.junit.jupiter.api.DisplayNameGeneration;
 import org.junit.jupiter.api.DisplayNameGenerator;
